Modified debug text behavior to improve frame rate
Through testing it was discovered that the method of displaying
debug text that was employed previously (displaying text to
terminal and using clearscreen to keep text at the top of the
terminal) was actually causing a significant drop if frame rate.

Removed all print commands and clearscreen command, instead
opting to displaying frame rate in the upper-left corner of the
graphical window.  This allows users to view the frame rate more
easily, and does not reduce frame rate in the manner that the
previous setup did.